requirements.txt
Flask
openai
python-dotenv
PyMuPDF